requirements.txt
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.5
attrs==19.3.0
autopep8==1.5
billiard==3.6.3.0
black==19.10b0
cachetools==4.0.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.41.0
django-decouple==2.1
django-redis-cache==2.1.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.17.0
idna==2.9
importlib-metadata==1.5.0
kombu==4.6.8
oauthlib==3.1.0
pathspec==0.7.0
protobuf==3.11.3
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
python3-openid==3.1.0
pytz==2019.3
redis==3.4.1
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
rope==0.16.0
rsa==4.0
selenium==3.141.0
sentry-sdk==0.14.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
whitenoise==5.0.1
zipp==3.1.0
